refactor: restructure pages — Recovery, Nutrition, Sidebar nav
- /data redirige vers /recovery (TrainingTab + RecoveryTab supprimés)
- Création /nutrition avec NutritionTab complet (suivi repas, macros, IA, plan)
- Sidebar : /data → /recovery "Récupération", ajout /nutrition "Nutrition"
